Expose error module to out of tree users
While attempting to add a new extension without separating it from it's host project, I
noticed I could not use any of the standard sbi-rs error types.

This is a fairly quick and dirty fix, but it does allow me to build extensions outside
of SBI-RS (while still using sbi-rs).