sbi, api: Use i64 in SbiReturn type
This moves the type to correctly match the specification. Through the
use of the From trait on SbiReturn it is possible to get access to the
result of the call as the desired type.

In all cases the value is interpretted as the raw bit value with no
runtime checking (e.g. no checking that a u64 fits into the i64.)

The individual users of the the ecall_send() API can choose the type
that they desire and interpret as required.

See: https://github.com/rivosinc/salus/issues/38

Signed-off-by: Rob Bradford <rbradford@rivosinc.com>